+++ Tuneful/Tuneful.swift
@@ -26,11 +26,12 @@ class AppDelegate: NSObject, NSApplicationDelegate, NSMenuDelegate {
         }
     }
     
+    // Windows
     private var onboardingWindow: OnboardingWindow!
     private var miniPlayerWindow: MiniPlayerWindow = MiniPlayerWindow()
-    private var popover: NSPopover!
     
     // Popover
+    private var popover: NSPopover!
     static let popoverWidth: CGFloat = 210
     
     // Status bar
@@ -41,6 +42,12 @@ class AppDelegate: NSObject, NSApplicationDelegate, NSMenuDelegate {
     private var playerManager: PlayerManager!
     private var statusBarItemManager: StatusBarItemManager!
     private var statusBarPlaybackManager: StatusBarPlaybackManager!
+    
+    private let updateController = SPUStandardUpdaterController(
+        startingUpdater: true,
+        updaterDelegate: nil,
+        userDriverDelegate: nil
+    )
     
     // Settings
     let GeneralSettingsViewController: () -> SettingsPane = {
@@ -235,10 +242,10 @@ class AppDelegate: NSObject, NSApplicationDelegate, NSMenuDelegate {
         
         let updates = NSMenuItem(
             title: "Check for updates...",
-            action: #selector(SUUpdater.checkForUpdates(_:)),
+            action: #selector(updateController.updater.checkForUpdates),
             keyEquivalent: ""
         )
-        updates.target = SUUpdater.shared()
+        updates.target = updateController.updater
         statusBarMenu.addItem(updates)
         
         statusBarMenu.addItem(.separator())
@@ -448,8 +455,8 @@ class AppDelegate: NSObject, NSApplicationDelegate, NSMenuDelegate {
         SettingsWindowController(
             panes: [
                 GeneralSettingsViewController(),
-                PopoverSettingsViewController(),
                 MenuBarSettingsViewController(),
+                PopoverSettingsViewController(),
                 MiniPlayerSettingsViewController(),
                 KeyboardShortcutsSettingsViewController(),
                 AboutSettingsViewController()
@@ -464,8 +471,8 @@ class AppDelegate: NSObject, NSApplicationDelegate, NSMenuDelegate {
         SettingsWindowController(
             panes: [
                 GeneralSettingsViewController(),
-                PopoverSettingsViewController(),
                 MenuBarSettingsViewController(),
+                PopoverSettingsViewController(),
                 MiniPlayerSettingsViewController(),
                 KeyboardShortcutsSettingsViewController(),
                 AboutSettingsViewController()